ml/pyproject.toml
[project]
name = "ml"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "mediapipe==0.10.14",
    "numpy<2.0.0",
    "opencv-python",
    "pandas",
    "psycopg2-binary",
    "python-dotenv",
    "python-multipart",
    "requests",
    "scikit-learn",
    "scipy",
    "fastdtw",
    "tensorflow-macos==2.15.0",
    "tqdm",
    "uvicorn",
    "yt-dlp",
]
